scoreKeeper/requirements.txt
Django>=3.1.1,<3.2
djangorestframework>=3.11.1,<3.12
gunicorn>=20.0.4,<20.1
psycopg2-binary>=2.8.5,<2.9
